Revert "Revert "add logs chat id""
This reverts commit afcda9f91023a2b2eb2ea626966ca2d922fc3487.

+++ src/main/kotlin/com/github/insanusmokrassar/TimingPostsTelegramBot/plugins/builtin/GarbageCollector.kt
@@ -66,7 +66,7 @@ class GarbageCollector(
         if (dataPair.second < config.minimalRate) {
             deletePost(
                 bot,
-                baseConfig.sourceChatId.toLong(),
+                baseConfig.sourceChatId,
                 dataPair.first
             )
         }
